API update: update stories.py
This commit is part of the work done on Lunii.PACKS to update the API to
match the one currently used in Lunii.QT.

Here the content of 'stories.py' is modifies in order to embed the
latest features on the stories managment. There is only one difference
with Lunii.QT: here the file 'unofficial.json' is not copied from QT
environment but from the project repo itself.